Fix the includes to contain relative paths
When using

    #include ""

all compilers first check in the current directory of the file that
contains the directive. Then they check in different orders in the
directories specified with -I. A natural order is to check in the order
of the specifications given (GCC does that). If we now have a compiler
command like

    CC ... -I/path/to/libA -I/path/to/LinAlg/include

and libA contains a header with the same name 'my_header.h'
as a header in LinAlg, LinAlg/include/part/other.h includes
libA/my_header.h instead of LinAlg/include/my_header.h.

Having relative paths in all headers in LinAlg/include fixes this
problem.