Add Sparse.reallocate(), [Dense|Sparse].is_on_X()
Sparse didn't have a reallocate(). This became necessary due to the
removal of the copy constructors and assignment operator.

The is_on_<X>() members allow the users to avoid #ifdefs in their code
as they are available irrespective of the compilation flags (as opposed
to the members of LinAlg::Location whose availability does depend on
the compilation flags).