Implement global CUDA streams
These streams can be reused for all transfers, avoiding creation of a
stream for each transfer. This primarily helps with tracing as each CUDA
stream shows up as one 'thread of execution' in a score-p trace and
vampir is limiting the number of threads that can be shown. It has
probably the lower overhead than the implementation we had before but
performance gains are marginal. So far it is preliminary in the sense
that only one device per rank is supported.